実装と定義を分ける意味が特にないため、discord.d.tsを削除
