+++ src/commands/cnfRmdBump.ts
@@ -1,8 +1,7 @@
 import { ChatInputCommandInteraction, Role, SlashCommandBuilder } from "discord.js";
-import { GetReplyEmbed, ReplyEmbedType } from "../services/discord";
+import { Command, GetReplyEmbed, ReplyEmbedType } from "../services/discord";
 import keyvs, { KeyvKeys } from "../services/keyvs";
 import { __t } from "../services/locale";
-import { Command } from "../types/discord";
 
 export const cnfRmdBumpCommand: Command = {
     data: new SlashCommandBuilder()
