+++ src/events/channelDelete.ts
@@ -1,8 +1,8 @@
 import { ChannelType, DMChannel, Events, GuildChannel, VoiceChannel } from "discord.js";
+import { BotEvent } from "../services/discord";
 import keyvs, { KeyvKeys, KeyvsError } from "../services/keyvs";
 import { __t } from "../services/locale";
 import { logger } from "../services/logger";
-import { BotEvent } from "../types/discord";
 
 export const channelDeleteEvent: BotEvent = {
     name: Events.ChannelDelete,
