+++ src/events/guildDelete.ts
@@ -1,8 +1,8 @@
 import { Events } from "discord.js";
+import { BotEvent } from "../services/discord";
 import keyvs from "../services/keyvs";
 import { __t } from "../services/locale";
 import { logger } from "../services/logger";
-import { BotEvent } from "../types/discord";
 
 export const guildDeleteEvent: BotEvent = {
     name: Events.GuildDelete,
