+++ src/events/intaractionCreate.ts
@@ -1,9 +1,8 @@
 import { Events, Interaction } from "discord.js";
-import { GetReplyEmbed, ReplyEmbedType } from "../services/discord";
+import { BotEvent, GetReplyEmbed, ReplyEmbedType } from "../services/discord";
 import keyvs, { KeyvsError } from "../services/keyvs";
 import { __t } from "../services/locale";
 import { logger } from "../services/logger";
-import { BotEvent } from "../types/discord";
 
 export const interactionCreateEvent: BotEvent = {
     name: Events.InteractionCreate,
