+++ src/events/ready.ts
@@ -1,8 +1,8 @@
 import { ActivityType, Client, Events } from "discord.js";
+import { BotEvent } from "../services/discord";
 import keyvs from "../services/keyvs";
 import { __t } from "../services/locale";
 import { logger } from "../services/logger";
-import { BotEvent } from "../types/discord";
 
 export const readyEvent: BotEvent = {
     name: Events.ClientReady,
