+++ src/events/voiceStateUpdate.ts
@@ -1,9 +1,8 @@
 import { ChannelType, Events, VoiceChannel, VoiceState } from "discord.js";
-import { GetReplyEmbed, ReplyEmbedType } from "../services/discord";
+import { BotEvent, GetReplyEmbed, ReplyEmbedType } from "../services/discord";
 import keyvs, { KeyvKeys, KeyvsError } from "../services/keyvs";
 import { __t } from "../services/locale";
 import { logger } from "../services/logger";
-import { BotEvent } from "../types/discord";
 
 export const voiceStateUpdateEvent: BotEvent = {
     name: Events.VoiceStateUpdate,
